// template/src/umbraco-api/api/generated/exampleApi.zod.ts
export const createItemBodyNameMax = 255;

export const createItemBodyDescriptionMax = 2000;

export const createItemBodyIsActiveDefault = true;

export const createItemBody = zod.object({
  "name": zod.string().min(1).max(createItemBodyNameMax).describe('Item name'),
  "description": zod.string().max(createItemBodyDescriptionMax).nullish().describe('Optional description'),
  "isActive": zod.boolean().default(createItemBodyIsActiveDefault).describe('Whether the item is active')
})